Update patch set 2
Patch Set 2:

Zack,   

Will this job runs on each code submission?   The project is in constant churn at this time, so it's not advisable to run these tests right now.    Is there a way to run these only manually for now in order to have less things to fix once the project reach a stable stable?

Thanks
Khen

Patch-set: 2
Label: Verified=0