Update patch set 1
Patch Set 1:

> imo it looks like we have too many "voltha-go-test-*" pipelines.
 > This test should be able to use one of the other pipelines we
 > already have in place ("voltha-bbsim-tests.groovy"). That pipeline
 > has an additional stage, "Build Images". So that stage should be
 > paramaterized with a boolean in this job and the other job that
 > uses that pipeline.
 > 
 > Something like - "when { expression { return params.BuildImages }
 > }" in the "Build Images" stage. See: https://gerrit.opencord.org/gitweb?p=ci-management.git;a=blob;f=jjb/pipeline/voltha-atest-provisioning.groovy;h=de7ef0b60baaf0991c6e0afc91bf71d8fcbfce19;hb=refs/heads/master#l55

I agree that this pipeline could be combined with voltha-bbsim-tests.groovy.  I'm reluctant to take that approach because lots of jobs depend on voltha-bbsim-tests.groovy, adding this pipeline's functionality would require a number of changes to that, and we don't really have a good way to test pipeline changes short of actually deploying them on Jenkins.  Since I'm not an expert in JJB I think there's a good chance that lots of jobs end up temporarily broken.  I think creating a  separate pipeline for testing helm charts is acceptable.

Patch-set: 1
Label: Code-Review=0