Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> What is the workflow for developing this ? Can it be tested based on an in review patch?

For this patch, you could commit it without the trigger, and then manually trigger it in Jenkins.

I can give you rights to edit the job in Jenkins, then when you have it all correct we can make a new patch that makes the needed changes and enables the trigger.

Patch-set: 6