Fix onos-app-release job to work in harmony with other jobs
This patch updates the legacy cord-onos-publisher/Jenkinsfile.release to
work with the new maven-* jobs. For consistency with outher job
pipelines, the pipeline  implementation is now hosted in ci-management,
instead of the cord-onos-publisher repo.

The new implementation updates the <version> tag in the root pom.xml for
the given app repo, and pushes two new Gerrit changes:
  (1) With version the given ${version} (e.g., 1.0.0)
  (2) With ${nextVersion}-SNAPSHOT (e.g., 1.1.0-SNAPSHOT)

Users must manually approve and merge these changes on Gerrit.
Differently from the previous implementation, once change (1) is merged,
it's up to the maven-publish and version-tag jobs to complete the
release by uploading artifacts to Sonatype and creating Git tags.

Change-Id: I495a37deda6fd13abf39e9e4ecd75d8899d60314