Update patch set 1
Patch Set 1:

(2 comments)

I've updated the pipeline as per our discussion on Slack. Not it should be compatible with the existing maven-* jobs. Please check the comment at the top for an explanation of the steps. I hope we can target some of your comments as a later task. Even if this file adds a new pipeline file, it is based on the legacy cord-onos-publisher/Jenkinsfile.release.

Patch-set: 1
Label: Code-Review=0