Update patch set 3
Patch Set 3:

I think the purpose of this pipeline is to bring up VOLTHA and run a test on it using a single (configurable) makeTarget.  As Zack suggests, one possibility would be to create a new job that runs the new makeTarget rather than editing the pipeline.  

If we want to create a single pipeline that runs a battery of different tests, possibly reconfiguring the system between each one, then it seems that we should rethink the entire pipeline.

Patch-set: 3