Update patch set 3
Patch Set 3:

My inputs: This patch set includes a temporary way to test rw-core scenario because of a known issue.  Once the issue is fixed then we will not require to teardown the entire system.  Also, there will be several other test cases that might require a re-install of the entire system and it's just the nature of the test.  In my opinion it's best to avoid creating multiple jobs.  Finally, for this scenario I feel it's still good to continue the way it's defined in this pipeline.

Patch-set: 3