Update patch set 2
Patch Set 2:

'release' is a string that is currently used in the name of the template.  It's values currently are `release` or `master` if you see our jobs, we tag them towards the end.  We could have used the same variable probably to pull helm charts accordingly instead of boolean 'released'.  I can work on a different patch for it later. For now, current patch is not affected

Patch-set: 2