Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If we just change NAME=minimal would the other changes be required? 

No they won't but given that those file names depends on the $NAME value I don't see what's the harm in using that.

> Also, might we want to default the NAME in `kind-voltha` to `minimal` as opposed to `local` for backward compatibility?

Is this a comment for the CI patch or the kind-voltha one? The name of the cluster we're creating for a single run of the tests on an ephemeral VM shouldn't pose any challenge to backward compatibility.

Patch-set: 1