Update patch set 2
Patch Set 2:

Ideally we should have a single "periodic" pipeline. I left a comment in gerrit.opencord.org/c/ci-management/+/24340/4/jjb/voltha-e2e.yaml (line 287). If we can change the parameters take a list of make targets instead of having different parameter it would be straight forward to do so in a loop.
If not we can always use a "when" conditional to trigger some steps or not, based on wether that make target is defined or not. Can you take a look on these approaches Monday?

Patch-set: 2
Reviewer: Gerrit User 1000007 <1000007@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>