Fix port-forward argument being null
This change addresses a problem where jenkins uses this command, causing
a parsing error in kubectl:

kubectl port-forward --address 0.0.0.0 -n infra svc/voltha-infra-etcd null:2379

This may have repurcussions further on when other tools such as voltctl
expect to find etcd at 127.0.0.1:2379.

This call is issued in voltha-dt-physical-functional-tests.groovy which
does not have a value for VolthaEtcdPort. The faulty call is wrapped in
a "while true" loop which masks the exit status of kubectl.

VolthaEtcdPort was initially introduced to have a different port (9999)
for the ONF demo pod, but the variable is now being set to override
the default port (2379) with 9999 for many other jobs.

It is unclear whether VolthaEtcdPort is still needed. It might be best
to remove the variable altogether, but as long as it is used, the value
must be set for all jobs using voltha-dt-physical-functional-tests.groovy
and match the value set for the build job that prepares the test.

For now, this change just tries to fix the resources needed for
config-pod berlin-community-pod-2-gpon-zyxel.

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>
Change-Id: Ib858a5d7d9b6b0cfe54bb4261eb68c5f3e96eabc