Abort if voltctl cannot find etcd
The voltctl commands we used to set log levels return only a rather
unhelpful error message ("context deadline exceeded") if voltctl
is unable to connect to etcd. The exit status in that case is 0,
wrongly indicating success.

This happens quite easily if kubectl was configured to forward etcd to
a port other than the one expected by voltctl.

Using "voltctl log level list" as we do with this change has two effects
if voltctl cannot find etcd. First, it exits with status 1 which should
abort the pipeline, alerting the user to a configuration problem. Second,
it prints a more helpful error message:

  Is ETCD available at localhost:2379?

Signed-off-by: Roger Luethi <roger.luethi@bisdn.de>
Change-Id: I7c2b0fdd8a2028c4b85b1a9b81c35dd194de99ca