Revert use of pgrep_port_forward in bbsim-tests
It looks like the pipeline now aborts when pgrep_port_forward() is
called. The log files do not contain any indications as to why the call
would fail, but looking at the code it might be because no arguments
are passed to the method which expects multiple arguments.

This is not a proper fix, but it may help us find out why the pipeline
aborts even earlier than it used to.

Change-Id: I73148ce71748cec4684f65ad774ee13abb45c2e1
Signed-off-by: Roger Luethi <roger@opennetworking.org>