Remove fixPerms() method
Now that the problem causing "Permission denied" log entries is
understood, the fixPerms() method is no longer needed. All we need to
do is fix all instances where code tries to execute the configuration
file instead of passing it to voltctl.

Also include an extra change in jjb/ to trigger jenkins job in vars/.

Signed-off-by: Roger Luethi <roger@opennetworking.org>
Change-Id: I979e86eeabbb2d1df38286db66754b125f991a5e