Make port-forwarding removal more robust
The pkill command in pkill_port_forward is wrapped by a pgrep guard
that looks for processes that need killing. The pgrep check is more
strict than the pkill command and prevents the execution of pkill
when it shouldn't.

We could simply drop the guard and handle the pkill exit status, but
we keep the guard for now and have it print information about matching
processes.

Instead of using "--count", rely on the pgrep exit status to indicate
if any matching processes were found.

While we're at it, fix the misleading label for the script.

Also include an extra change in jjb/ to trigger jenkins job in vars/.

Signed-off-by: Roger Luethi <roger@opennetworking.org>
Change-Id: I99158e5d6477091a2f9bf6d0c8431c0ca53ef1b2