CI: Remove ci-man jobs from JJB, change GHA WF
The previous GHA workflow did not include voting, so this switches to
one that does.

Packer jobs will remain in JJB for now, as they are not yet properly
implemented for AWS-based Jenkins clouds. The merge job also remains
in place for now, though it will be replaced once GHA jobs are
functioning correctly.

This also includes some yamllint fixes to .pre-commit-config.yaml.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>
Change-Id: Ife44a2cfed9c5a5bc4ca8a2ac3f38f9a95210d8d