Fix: Correct issues with github-release.sh
The github-release shell script was redesigned while CI was down,
which introduced some broken code. In particular, references to other
files in the tree were added, but these don't work in a Jenkins job
where the shell script becomes part of the job, separated from the
ci-management file tree.

The job has been adapted to run both locally with these library files,
and in a CI environment.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>
Change-Id: I22a52a3ac1f1cbde1c3990f5bbcff9412d3865ff