Remove .kube/config in k8s cluster setup
In some newer systems, we're seeing .kube/config as an existing file
that is linked to /var/lib/kubelet/kubeconfig, which presents issues
when trying to write a new config to that path. Since this is intended
to fully overwrite what's there, it's safe to remove it first,
whether it's a file or a symlink.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>
Change-Id: I8edb3897f3f7649ac085528541b053b557526792